Run EMR step within data pipeline

+++ bin/copy_env.sh
@@ -32,10 +32,26 @@ ask_to_confirm () {
     done
 }
 
-ask_to_confirm "Are you sure you want to overwrite '$CLUSTER_ENVIRONMENT?'"
+ask_to_confirm "Are you sure you want to overwrite '$CLUSTER_ENVIRONMENT' (from '$CLUSTER_SOURCE_ENVIRONMENT')?"
 
 set -e -x
 
-for FOLDER in config jars bootstrap; do
-    aws s3 cp --recursive "s3://$CLUSTER_BUCKET/$CLUSTER_SOURCE_ENVIRONMENT/$FOLDER/" "s3://$CLUSTER_BUCKET/$CLUSTER_ENVIRONMENT/$FOLDER/"
-done
+if [[ "$CLUSTER_SOURCE_ENVIRONMENT" = "local" ]]; then
+    aws s3 cp "bin/bootstrap.sh" "s3://$CLUSTER_BUCKET/$CLUSTER_ENVIRONMENT/bootstrap/"
+    for FILE in config/*; do
+        aws s3 cp "$FILE" "s3://$CLUSTER_BUCKET/$CLUSTER_ENVIRONMENT/config/"
+    done
+    for FILE in python/requirements.txt \
+                python/dist/redshift-etl-0.6.4.tar.gz \
+                jars/commons-csv-1.4.jar \
+                jars/postgresql-9.4.1208.jar \
+                jars/RedshiftJDBC41-1.1.10.1010.jar \
+                jars/spark-csv_2.10-1.4.0.jar
+    do
+        aws s3 cp "$FILE" "s3://$CLUSTER_BUCKET/$CLUSTER_ENVIRONMENT/jars/"
+    done
+else
+    for FOLDER in config jars bootstrap; do
+        aws s3 cp --recursive "s3://$CLUSTER_BUCKET/$CLUSTER_SOURCE_ENVIRONMENT/$FOLDER/" "s3://$CLUSTER_BUCKET/$CLUSTER_ENVIRONMENT/$FOLDER/"
+    done
+fi
